Chore: Refactor Repo for Quality of Life
