+++ main.py
@@ -1,8 +1,8 @@
-from modules.simulation import Simulation
+from modules.environment.simulation import Simulation
 if __name__ == "__main__":    
     N_SIMS = 10
     i = 0
-    s = Simulation(strategy='random')
+    s = Simulation(strategy='random', render='human')
     while i < N_SIMS:
         print(f"Simulation {i}/{N_SIMS}")
         print(s.run())
