+++ modules/actors/__init__.py
@@ -1,0 +1,3 @@
+from modules.actors.constants import action_codes
+import modules.actors.adversaries as adversaries
+import modules.actors.agents as agents
