+++ modules/actors/adversaries/__init__.py
@@ -1,0 +1,3 @@
+from modules.actors.adversaries.base_adversary import BaseAdversary
+from modules.actors.adversaries.greedy_adversary import GreedyAdversary
+from modules.actors.adversaries.random_adversary import RandomAdversary
