+++ modules/actors/agents/__init__.py
@@ -1,0 +1,3 @@
+from modules.actors.agents.base_agent import BaseAgent
+
+from modules.actors.agents.random_agent import RandomAgent
